% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHMassBank-package.R
\docType{package}
\name{AHMassBank-package}
\alias{AHMassBank}
\alias{AHMassBank-package}
\title{AHMassBank: MassBank Annotation Resources for AnnotationHub}
\description{
Supplies AnnotationHub with MassBank metabolite/compound annotations bundled in CompDb SQLite databases. CompDb SQLite databases contain general compound annotation as well as fragment spectra representing fragmentation patterns of compounds' ions. MassBank data is retrieved from https://massbank.eu/MassBank and processed using helper functions from the CompoundDb Bioconductor package into redistributable SQLite databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jorainer/AHMassBank}
  \item Report bugs at \url{https://github.com/jorainer/AHMassBank/issues}
}

}
\author{
\strong{Maintainer}: Johannes Rainer \email{Johannes.Rainer@eurac.edu} (\href{https://orcid.org/0000-0002-6977-7147}{ORCID})

}
\keyword{internal}
